+++ Configuration/TCA/Overrides/sys_file_reference.php
@@ -22,11 +22,43 @@ declare(strict_types=1);
  */
 
 use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 defined('TYPO3') or die();
 
 $extensionConfiguration = GeneralUtility::makeInstance(ExtensionConfiguration::class);
+
+ExtensionManagementUtility::addTCAcolumns(
+    'sys_file_reference',
+    [
+        'tx_mindfula11y_decorative' => [
+            'exclude' => true,
+            'label' => 'LLL:EXT:mindfula11y/Resources/Private/Language/Database.xlf:sysFileReference.columns.mindfula11y.decorative',
+            'description' => 'LLL:EXT:mindfula11y/Resources/Private/Language/Database.xlf:sysFileReference.columns.mindfula11y.decorative.description',
+            'config' => [
+                'type' => 'check',
+                'renderType' => 'checkboxToggle',
+                'default' => 0,
+            ],
+        ],
+    ]
+);
+
+ExtensionManagementUtility::addFieldsToPalette(
+    'sys_file_reference',
+    'imageoverlayPalette',
+    'tx_mindfula11y_decorative',
+    'before:alternative'
+);
+
+$decorativeDisplayCondition = 'FIELD:tx_mindfula11y_decorative:!=:1';
+foreach (['alternative', 'title'] as $fieldName) {
+    $existingDisplayCondition = $GLOBALS['TCA']['sys_file_reference']['columns'][$fieldName]['displayCond'] ?? null;
+    $GLOBALS['TCA']['sys_file_reference']['columns'][$fieldName]['displayCond'] = empty($existingDisplayCondition)
+        ? $decorativeDisplayCondition
+        : ['AND' => [$existingDisplayCondition, $decorativeDisplayCondition]];
+}
 
 if (
     !$extensionConfiguration->get('mindfula11y', 'disableAltTextGeneration') &&