+++ ext_tables.sql
@@ -1,0 +1,3 @@
+CREATE TABLE sys_file_reference (
+	tx_mindfula11y_decorative smallint(5) unsigned DEFAULT '0' NOT NULL
+);